0.9-dev-61

+++ src/overlay.c
@@ -60,25 +60,25 @@ int drawCrossHairs(bsr_config_t *bsr_config, bsr_state_t *bsr_state) {
   // draw crosshairs
   //
   for (i=(half_res_x - (res_y * 0.02)); i < (half_res_x - (res_y * 0.005)); i++) {
-    image_composition_p=image_buf + (int)(res_x * half_res_y) + i;
+    image_composition_p=image_buf + ((int)res_x * (int)half_res_y) + i;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=(half_res_x + (res_y * 0.005)); i < (half_res_x + (res_y * 0.02)); i++) {
-    image_composition_p=image_buf + (int)(res_x * half_res_y) + i;
+    image_composition_p=image_buf + ((int)res_x * (int)half_res_y) + i;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=(half_res_y - (res_y * 0.02)); i < (half_res_y - (res_y * 0.005)); i++) {
-    image_composition_p=image_buf + (int)(res_x * (double)i) + (int)half_res_x;
+    image_composition_p=image_buf + ((int)res_x * i) + (int)half_res_x;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=(half_res_y + (res_y * 0.005)); i < (half_res_y + (res_y * 0.02)); i++) {
-    image_composition_p=image_buf + (int)(res_x * (double)i) + (int)half_res_x;
+    image_composition_p=image_buf + ((int)res_x * i) + (int)half_res_x;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
@@ -108,37 +108,37 @@ int drawGridLines(bsr_config_t *bsr_config, bsr_state_t *bsr_state) {
   // draw select raster lines
   //
   for (i=0; i < res_x; i++) {
-    image_composition_p=image_buf + (int)(res_x * (res_y * 0.25)) + i;
+    image_composition_p=image_buf + (int)res_x * (int)(res_y * 0.25) + i;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=0; i < res_x; i++) {
-    image_composition_p=image_buf + (int)(res_x * half_res_y) + i;
+    image_composition_p=image_buf + (int)res_x * (int)half_res_y + i;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=0; i < res_x; i++) {
-    image_composition_p=image_buf + (int)(res_x * (res_y * 0.75)) + i;
+    image_composition_p=image_buf + (int)res_x * (int)(res_y * 0.75) + i;
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=0; i < res_y; i++) {
-    image_composition_p=image_buf + (int)(res_x * (double)i) + (int)(res_x * 0.25);
+    image_composition_p=image_buf + (int)res_x * i + (int)(res_x * 0.25);
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=0; i < res_y; i++) {
-    image_composition_p=image_buf + (int)(res_x * (double)i) + (int)(half_res_x);
+    image_composition_p=image_buf + (int)res_x * i + (int)(half_res_x);
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
   }
   for (i=0; i < res_y; i++) {
-    image_composition_p=image_buf + (int)(res_x * (double)i) + (int)(res_x * 0.75);
+    image_composition_p=image_buf + (int)res_x * i + (int)(res_x * 0.75);
     image_composition_p->r=0.9;
     image_composition_p->g=0.0;
     image_composition_p->b=0.0;
